@@ -11,47 +11,54 @@
   "license": "MIT",
   "scripts": {
     "test": "jest",
-    "dev": "webpack-dev-server --config ./build/webpack.dev.config.js",
+    "dev": "webpack serve --config ./build/webpack.dev.config.js",
     "build": "rm -rf dist && webpack --config ./build/webpack.prod.config.js"
   },
   "dependencies": {
-    "@babel/runtime": "^7.10.4"
+    "@babel/runtime": "^7.14.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.10.4",
-    "@babel/plugin-transform-modules-commonjs": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@types/jest": "^26.0.3",
-    "@types/jquery": "^3.5.0",
-    "autoprefixer": "^9.8.4",
-    "babel-jest": "^26.1.0",
-    "babel-loader": "^8.1.0",
+    "@babel/core": "^7.14.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.14.2",
+    "@babel/plugin-transform-modules-commonjs": "^7.14.0",
+    "@babel/plugin-transform-runtime": "^7.14.3",
+    "@babel/preset-env": "^7.14.2",
+    "@types/jest": "^26.0.23",
+    "@types/jquery": "^3.5.5",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
+    "autoprefixer": "^10.2.5",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
     "cache-loader": "^4.1.0",
-    "copy-webpack-plugin": "^6.0.3",
-    "css-loader": "^3.6.0",
-    "cssnano": "^4.1.10",
+    "copy-webpack-plugin": "^8.1.1",
+    "css-loader": "^5.2.4",
+    "cssnano": "^5.0.3",
+    "eslint": "^7.26.0",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "fibers": "^5.0.0",
-    "file-loader": "^6.0.0",
-    "html-webpack-plugin": "^4.3.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.3.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.1.0",
-    "jquery": "^3.5.1",
-    "mini-css-extract-plugin": "^0.9.0",
-    "mock-xmlhttprequest": "^6.1.0",
-    "postcss-loader": "^3.0.0",
-    "postcss-normalize": "^9.0.0",
-    "sass": "^1.26.9",
-    "sass-loader": "^8.0.2",
-    "style-loader": "^1.2.1",
-    "ts-jest": "^26.1.1",
-    "ts-loader": "^7.0.5",
-    "typescript": "^3.9.5",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.2"
+    "jest": "^26.6.3",
+    "jquery": "^3.6.0",
+    "mini-css-extract-plugin": "^1.6.0",
+    "mock-xmlhttprequest": "^7.0.4",
+    "postcss": "^8.2.15",
+    "postcss-loader": "^5.3.0",
+    "postcss-normalize": "^10.0.0",
+    "raw-loader": "^4.0.2",
+    "sass": "^1.32.13",
+    "sass-loader": "^11.1.1",
+    "style-loader": "^2.0.0",
+    "ts-jest": "^26.5.6",
+    "ts-loader": "^9.2.1",
+    "typescript": "^4.2.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.37.1",
+    "webpack-cli": "^4.7.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3"
   }
 }
