@@ -1,16 +1,1 @@
-
-module.exports = {
-    plugins: [
-        require('autoprefixer')(),
-        require('postcss-normalize')(),
-        require('cssnano')({
-            preset: [
-                'default', {
-                    discardComments: {
-                        removeAll: true
-                    }
-                }
-            ]
-        })
-    ]
-}
+config/postcss.config.js
